This is the proof of concept for Carbon 14.  It is just good enough to run a classic PowerPC build of the "Traffic Light" DTS sample as a native, Carbonized app on Mac OS X.
The only modification to "Traffic Light" is the addition of a 'carb'
resource (to fool OS X into launching it), and maybe a property list.

There are three anomolies.  First, the region used to adjust the
cursor is off; Traffic Light seems to assume a B&W port by accessing
portBits.bounds (???).  Second, the "about" box text is slightly
clipped.  Third, the "Quit" menu item is of course under the "File"
menu, and the OS X quit item doesn't work.